pal: Add palRverot
Changed the data types from SLA to use double instead of float.

Also added the RV tests from SLA.

+++ libraries/pal/pal.h
@@ -465,7 +465,7 @@ void palRefv ( double vu[3], double refa, double refb, double vr[3] );
 
 void palRefz ( double zu, double refa, double refb, double *zr );
 
-float palRverot ( float phi, float ra, float da, float st );
+double palRverot ( double phi, double ra, double da, double st );
 
 double palRvgalc ( double r2000, double d2000 );
